tk - refactored two classes out of entities
